moved Command to Command package
