+++ src/main/java/com/spbgti/dispatcherapp/Entity/Event/Command/UpdateCommand.java
@@ -1,4 +1,6 @@
-package com.spbgti.dispatcherapp.Entity.Event;
+package com.spbgti.dispatcherapp.Entity.Event.Command;
+
+import com.spbgti.dispatcherapp.Entity.Event.ClassParser;
 
 import javax.persistence.EntityManager;
 import java.lang.reflect.InvocationTargetException;
