+++ src/main/java/com/spbgti/dispatcherapp/Entity/Event/ModifyEvent.java
@@ -1,5 +1,6 @@
 package com.spbgti.dispatcherapp.Entity.Event;
 
+import com.spbgti.dispatcherapp.Entity.Event.Command.Command;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 import javax.persistence.EntityManager;
